added activities folder and moved all activites to that folder
