+++ app/src/main/AndroidManifest.xml
@@ -13,16 +13,16 @@
         android:theme="@style/Theme.MediHub"
         tools:targetApi="31">
         <activity
-            android:name=".AdminActivity"
+            android:name=".activities.AdminActivity"
             android:exported="false" />
         <activity
-            android:name=".DoctorActivity"
+            android:name=".activities.DoctorActivity"
             android:exported="false" />
         <activity
-            android:name=".PatientActivity"
+            android:name=".activities.PatientActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -30,11 +30,11 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".RegistrationOptionActivity" />
-        <activity android:name=".PatientRegistrationActivity" />
-        <activity android:name=".DoctorRegistrationActivity" />
-        <activity android:name=".LoginActivity" />
-        <activity android:name=".WelcomeActivity" />
+        <activity android:name=".activities.RegistrationOptionActivity" />
+        <activity android:name=".activities.PatientRegistrationActivity" />
+        <activity android:name=".activities.DoctorRegistrationActivity" />
+        <activity android:name=".activities.LoginActivity" />
+        <activity android:name=".activities.WelcomeActivity" />
     </application>
 
 </manifest>
