+++ app/src/main/java/com/example/medihub/activities/DoctorActivity.java
@@ -1,4 +1,4 @@
-package com.example.medihub;
+package com.example.medihub.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -7,6 +7,7 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 
+import com.example.medihub.R;
 import com.example.medihub.models.DoctorProfile;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.database.FirebaseDatabase;
