+++ app/src/main/java/com/example/medihub/activities/PatientRegistrationActivity.java
@@ -1,4 +1,4 @@
-package com.example.medihub;
+package com.example.medihub.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -6,15 +6,13 @@ import android.util.Log;
 import android.util.Patterns;
 import android.view.View;
 import android.widget.Button;
-import android.widget.CheckBox;
 import android.widget.EditText;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.example.medihub.enums.UserRole;
-import com.example.medihub.models.DoctorProfile;
+import com.example.medihub.R;
 import com.example.medihub.models.PatientProfile;
 import com.example.medihub.models.UserProfile;
 import com.google.android.gms.tasks.OnCompleteListener;
@@ -25,9 +23,7 @@ import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.Map;
 
 public class PatientRegistrationActivity extends AppCompatActivity  {
