+++ app/src/main/java/com/example/medihub/activities/RegistrationOptionActivity.java
@@ -1,12 +1,13 @@
-package com.example.medihub;
+package com.example.medihub.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
+
+import com.example.medihub.R;
 
 public class RegistrationOptionActivity extends AppCompatActivity {
     @Override
