+++ app/src/main/java/com/example/medihub/activities/WelcomeActivity.java
@@ -1,9 +1,8 @@
-package com.example.medihub;
+package com.example.medihub.activities;
 
 import android.os.Bundle;
 import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
-import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
@@ -11,6 +10,9 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.example.medihub.R;
+import com.example.medihub.models.DoctorProfile;
+import com.example.medihub.models.PatientProfile;
+import com.example.medihub.models.UserProfile;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DataSnapshot;
@@ -28,6 +30,7 @@ public class WelcomeActivity extends AppCompatActivity {
     private FirebaseDatabase firebaseDB;
     private Button homeButton;
     private Intent homeIntent = null;
+    private UserProfile userProfile;
 
 
     @Override
@@ -62,12 +65,15 @@ public class WelcomeActivity extends AppCompatActivity {
                     switch (role) {
                         case "admin":
                             homeIntent = new Intent(WelcomeActivity.this, AdminActivity.class);
+                            userProfile = dataSnapshot.getValue(UserProfile.class);
                             break;
                         case "patient":
                             homeIntent = new Intent(WelcomeActivity.this, PatientActivity.class);
+                            userProfile = dataSnapshot.getValue(PatientProfile.class);
                             break;
                         case "doctor":
                             homeIntent = new Intent(WelcomeActivity.this, DoctorActivity.class);
+                            userProfile = dataSnapshot.getValue(DoctorProfile.class);
                             break;
                     }
                 }
@@ -98,9 +104,10 @@ public class WelcomeActivity extends AppCompatActivity {
             @Override
             public void onClick(View view) {
                 if (homeIntent != null) {
+                    homeIntent.putExtra("user", userProfile);
                     startActivity(homeIntent);
                 } else {
-                    Toast.makeText(WelcomeActivity.this, "Redirect failed, please try again later.", Toast.LENGTH_LONG);
+                    Toast.makeText(WelcomeActivity.this, "Redirect failed, please try again later.", Toast.LENGTH_LONG).show();
                 }
             }
         });
