+++ app/src/main/res/layout/activity_admin.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AdminActivity">
+    tools:context=".activities.AdminActivity">
 
     <ScrollView
         android:layout_width="match_parent"
