+++ app/src/main/res/layout/activity_doctor.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DoctorActivity">
+    tools:context=".activities.DoctorActivity">
 
     <ScrollView
         android:layout_width="match_parent"
