+++ app/src/main/res/layout/activity_patient.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PatientActivity">
+    tools:context=".activities.PatientActivity">
 
     <ScrollView
         android:layout_width="match_parent"
