+++ app/src/main/res/layout/activity_welcome.xml
@@ -21,10 +21,10 @@
 
     <TextView
         android:id="@+id/textWelcome"
-        android:layout_width="match_parent"
-        android:layout_height="28dp"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
         android:layout_marginTop="50dp"
-        android:gravity="center"
+        android:gravity="center_horizontal"
         android:text="Welcome NAME"
         android:textSize="20sp"
         app:layout_constraintBottom_toTopOf="@+id/homeButton"
